refactor(slice4b): final-review nits — hoist renderAncestorContext, simplify reviewerFeedback, clamp maxRoundTrips>=1, template literal
- DRY: shared render-ancestor-context.ts (was duplicated in planner + reviewer);
  drop now-unused ContextPath imports.
- Simplify reviewerFeedback: verdict.feedback (the ternary was inside if(!verdict.pass)).
- Clamp maxRoundTrips to >=1 so a 0 config doesn't always exhaust before any work.
- Test: template literal for the clarify-marker assertion (clears Biome info).
